Add conditional command substitution and update demo analysis docs
- Implement conditional iidy command path substitution to avoid replacing when PATH iidy matches current executable
- Add is_iidy_on_path_same_as_current_exe function with canonical path comparison
- Update demo analysis documentation with command path substitution feature details
- Add comprehensive unit tests for command substitution logic

+++ iidy/src/cli.rs
@@ -634,8 +634,8 @@ pub struct GetImportArgs {
 #[derive(Args, Debug, Clone)]
 pub struct DemoArgs {
     pub demoscript: String,
-    #[arg(long, default_value_t = 1)]
-    pub timescaling: u32,
+    #[arg(long, default_value_t = 1.0)]
+    pub timescaling: f64,
 }
 
 #[derive(Args, Debug, Clone)]
